Updated CSS
Updated Navbar CSS and appropriate JSX to reflect that. Removed the text "Category:" Now it simply just says the category next to the ID in a cute little box.

I also made the Navbar look nicer and added some fluff to the buttons.

Made it so the log in warning fades in from the top and added a "!" after the username when logged in to better distinguish the user name from the rest of the bar.